fix: move model file
