+++ api/model/category.go
@@ -1,4 +1,4 @@
-package entity
+package model
 
 import (
 	"gorm.io/gorm"
